Pull out ServiceMethod base class
This will allow us to support multiple forms of invocation beyond HTTP calls.
